added example for function calling and made the required changes.
