+++ pyproject.toml
@@ -76,6 +76,7 @@ soundfile = [ "soundfile~=0.12.1" ]
 together = [ "openai~=1.59.0" ]
 websocket = [ "websockets~=13.1", "fastapi~=0.115.0" ]
 whisper = [ "faster-whisper~=1.1.0" ]
+openrouter = [ "openai~=1.59.0" ]
 
 [tool.setuptools.packages.find]
 # All the following settings are optional:
